add user table

+++ api/account_test.go
@@ -329,7 +329,7 @@ func TestListAccountsAPI(t *testing.T) {
 
 func randomAccount() db.Account {
 	return db.Account{
-		ID:       util.RandomInt(1, 1000),
+		ID:       util.RandomInt64(1, 1000),
 		Owner:    util.RandomOwner(),
 		Balance:  util.RandomMoney(),
 		Currency: util.RandomCurrency(),
